requirements.txt
fastapi==0.104.1
psycopg2-binary==2.9.7
pydantic>=2.5.3
uvicorn==0.24.0
python-multipart==0.0.6